pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ravah"
version = "0.1.0"
description = "Turn what you ship into a week of social content."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["cli", "content", "social-media", "ai", "gemini", "founder"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "typer[all]>=0.12.3",
    "rich>=13.7.1",
    "google-genai>=1.16.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.1",
]

[project.urls]
Homepage   = "https://github.com/muhtalhakhan/Ravah-App-OpenSouce"
Repository = "https://github.com/muhtalhakhan/Ravah-App-OpenSouce"
Issues     = "https://github.com/muhtalhakhan/Ravah-App-OpenSouce/issues"

[project.scripts]
ravah = "ravah.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["ravah"]
